# Use `parallel` library on builds