--- a/compiler/mimsa.cabal
+++ b/compiler/mimsa.cabal
@@ -228,7 +228,7 @@ library
   hs-source-dirs:
       src
   default-extensions:
-      Strict
+      StrictData
   build-depends:
       QuickCheck
     , aeson
@@ -248,6 +248,7 @@ library
     , monad-logger
     , mtl
     , openapi3
+    , parallel
     , parser-combinators
     , prettyprinter
     , servant
@@ -282,6 +283,8 @@ executable mimsa
       Paths_mimsa
   hs-source-dirs:
       repl
+  default-extensions:
+      StrictData
   ghc-options: -threaded -rtsopts -with-rtsopts=-N
   build-depends:
       QuickCheck
@@ -305,6 +308,7 @@ executable mimsa
     , mtl
     , openapi3
     , optparse-applicative
+    , parallel
     , parser-combinators
     , prettyprinter
     , servant
@@ -348,6 +352,8 @@ executable mimsa-server
       Paths_mimsa
   hs-source-dirs:
       server
+  default-extensions:
+      StrictData
   ghc-options: -threaded -rtsopts -with-rtsopts=-T
   build-depends:
       QuickCheck
@@ -371,6 +377,7 @@ executable mimsa-server
     , mtl
     , openapi3
     , optparse-applicative
+    , parallel
     , parser-combinators
     , prettyprinter
     , prometheus-client
@@ -481,6 +488,7 @@ test-suite mimsa-test
     , monad-logger
     , mtl
     , openapi3
+    , parallel
     , parser-combinators
     , prettyprinter
     , servant
@@ -520,6 +528,7 @@ benchmark mimsa-benchmark
     , monad-logger
     , mtl
     , openapi3
+    , parallel
     , parser-combinators
     , prettyprinter
     , servant